split signal tests
